requirements.txt
fastapi==0.95.1
uvicorn==0.21.1
sqlalchemy==2.0.9
PyMySQL==1.1.1
python-dotenv==1.0.0
pydantic==1.10.7
cryptography==42.0.4